fix: 주문 금액에 따른 계산 관련 메소드 위치 변경
