+++ src/main/java/order/model/Orders.java
@@ -47,4 +47,19 @@ public class Orders {
         }
         return count;
     }
+
+    public int deliverPrice() {
+        int totalOrderPrice = totalOrderPrice();
+        if (totalOrderPrice < 50000) {
+            return 2000;
+        }
+        if (totalOrderPrice < 100000) {
+            return 1000;
+        }
+        return 0;
+    }
+
+    public int totalPrice() {
+        return deliverPrice() + totalOrderPrice();
+    }
 }
